refactor: migrate to single export
