--- packages/core/src/components/button/button-variants.ts
@@ -1,10 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 Siemens AG
- *
- * SPDX-License-Identifier: MIT
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-export type Buttons = 'Primary' | 'Secondary';
