--- packages/core/src/components/datetime-picker/event.ts
@@ -1,14 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 Siemens AG
- *
- * SPDX-License-Identifier: MIT
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-export interface DateTimeSelectEvent {
-  from: string;
-  to: string;
-  time: string;
-}
