--- packages/core/src/components/dropdown/dropdown-trigger-event.tsx
@@ -1,13 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 Siemens AG
- *
- * SPDX-License-Identifier: MIT
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-/**
- * @internal
- */
-export type DropdownTriggerEvent = 'click' | 'hover' | 'focus';
