--- packages/core/src/exports.ts
@@ -1,18 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 Siemens AG
- *
- * SPDX-License-Identifier: MIT
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-export * from './components/date-picker/events';
-export * from './components/datetime-picker/event';
-export * from './components/dropdown/dropdown-trigger-event';
-export * from './components/modal/modal-utils';
-export * from './components/toast/toast-utils';
-export * from './components/tree-item/default-tree-item';
-export * from './components/tree/tree-model';
-export * from './components/upload/upload-file-state';
-export * from './components/utils/theme-switcher';
