+++ packages/core/src/index.ts
@@ -7,18 +7,27 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-export { Components, JSX } from './components';
-export { Buttons } from './components/button/button-variants';
-export { FilterState } from './components/category-filter/filter-state';
-export { InputState } from './components/category-filter/input-state';
-export { LogicalFilterOperator } from './components/category-filter/logical-filter-operator';
-export { DropdownTriggerEvent } from './components/dropdown/dropdown-trigger-event';
-export { FlipTileState } from './components/flip-tile/flip-tile-state';
+export * from './components';
+/**
+ * Category-filter uses classes instead of types
+ */
+// export { FilterState } from './components/category-filter/filter-state';
+// export { InputState } from './components/category-filter/input-state';
+// export { LogicalFilterOperator } from './components/category-filter/logical-filter-operator';
+/**
+ *
+ */
+// export { FlipTileState } from './components/flip-tile/flip-tile-state';
 export { ModalContainer } from './components/modal-container/modal-container';
 export { Modal } from './components/modal/modal';
+export * from './components/modal/modal-utils';
+export * from './components/toast/toast-utils';
+export * from './components/tree-item/default-tree-item';
+export * from './components/tree/tree-model';
+export * from './components/upload/upload-file-state';
 export { NotificationColor } from './components/utils/notification-color';
 export {
   convertToAbbreviationString,
   convertToRemString,
 } from './components/utils/rwd.util';
-export * from './exports';
+export * from './components/utils/theme-switcher';
